Adding routes and get functionality.
Added a route and a controller for each table in the database.

Currently only contains a get request, this returns all the information from a given table.

New POST/GET/PUT/DELETE requests will be put in the controller and to use them, they will be called in the route by declaring the function name and passing it to the route.